requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
openai
python-dotenv
pymupdf
